# docs/specs/openapi.yaml
openapi: 3.0.3
info:
  title: AI 맞춤형 운동프로그램 웹앱 API
  version: 1.0.0
  description: >
    기록 기반 폐루프 오토레귤레이션 코칭 웹앱의 백엔드 API.
    웹 우선(Next.js PWA) 구성: 인증은 httpOnly 세션 쿠키 + CSRF 헤더,
    결제는 국내 PG 정기결제(빌링키). 오프라인 로깅은 /sync 로 배치 동기화.
    규칙 버전(rules_version)과 reason_code 는 추천 재현/분석에 사용.
servers:
  - url: https://api.afc.example/v1
    description: production
  - url: https://staging-api.afc.example/v1
    description: staging
...
security:
  - cookieAuth: []
paths:
  /auth/owner/bootstrap:
    post:
      tags: [auth]
      summary: 단일 소유자 최초 등록(프로필·동의·admin·세션을 원자 생성)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OwnerBootstrapRequest' }
      responses:
        '201':
          description: '등록 성공. Set-Cookie sid=...; HttpOnly; Secure; SameSite=Lax'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/owner/login:
    post:
      tags: [auth]
      summary: 소유자 복구 코드로 재로그인(쿠키 삭제·기기 변경 대응)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OwnerLoginRequest' }
      responses:
        '200':
          description: '로그인 성공. Set-Cookie sid=...; HttpOnly; Secure; SameSite=Lax'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: 세션 갱신(refresh 회전)
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '204': { description: 갱신됨(Set-Cookie 재발급) }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [auth]
      summary: 로그아웃(세션 무효화)
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '204': { description: 로그아웃됨 }
  /me:
    get:
      tags: [users]
      summary: 내 프로필 조회
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [users]
      summary: 프로필 수정(목표 변경 시 프로그램 재구성 트리거)
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '400': { $ref: '#/components/responses/BadRequest' }
    delete:
      tags: [users]
      summary: 계정/데이터 삭제 요청(소프트 삭제 후 퍼지)
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '202': { description: 삭제 요청 접수(비동기 퍼지) }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/consents:
    post:
      tags: [users]
      summary: 개인정보/약관 동의 기록(PIPA)
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: '#/components/schemas/Consent' }
      responses:
        '204': { description: 기록됨 }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/export:
    get:
      tags: [users]
      summary: 내 데이터 즉시 JSON 내보내기(기계 판독)
      responses:
        '200':
          description: 현재 보유한 프로필·동의·운동 기록·파생 지표·접근 감사의 JSON 스냅샷
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataExport' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/calibration:
    post:
      tags: [users]
      summary: 간이 캘리브레이션(시작 무게 추정)
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                method: { type: string, enum: [survey, amrap] }
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      exercise_id: { type: string }
                      perceived_difficulty: { type: string, enum: [easy, moderate, hard] }
                      amrap_weight: { type: number, nullable: true }
                      amrap_reps: { type: integer, nullable: true }
      responses:
        '200':
          description: 초기 e1RM 산출
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Estimated1RM' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: httpOnly·Secure·SameSite=Lax 세션 쿠키. 변경 요청에는 X-CSRF-Token 헤더가 필요하다.
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
      description: CSRF 방어 토큰(쿠키 세션 기반 변경 요청 필수).
...
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: 인증 필요/만료
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: 현재 상태와 충돌
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: { type: object, additionalProperties: true, nullable: true }
    OwnerLoginRequest:
      type: object
      required: [owner_code]
      properties:
        owner_code:
          type: string
          minLength: 24
          maxLength: 256
          description: Secret Manager에만 보관하는 소유자 복구 코드. 요청·응답·DB·로그에 남기지 않는다.
    OwnerProfile:
      type: object
      required: [sex, birth_year, height_cm, weight_kg, body_fat_pct, goal, experience_level]
      properties:
        sex: { type: string, enum: [male, female, other] }
        birth_year: { type: integer, minimum: 1900, maximum: 2026 }
        height_cm: { type: number, minimum: 100, maximum: 250 }
        weight_kg: { type: number, minimum: 30, maximum: 300 }
        body_fat_pct: { type: number, nullable: true, minimum: 1, maximum: 70 }
        goal: { type: string, enum: [diet, hypertrophy, strength] }
        experience_level: { type: string, enum: [beginner, intermediate, advanced] }
    OwnerBootstrapRequest:
      allOf:
        - $ref: '#/components/schemas/OwnerLoginRequest'
        - type: object
          required: [profile, consents]
          properties:
            profile: { $ref: '#/components/schemas/OwnerProfile' }
            consents:
              type: array
              minItems: 1
              items: { $ref: '#/components/schemas/Consent' }
    AuthResult:
      type: object
      required: [csrf_token, user]
      properties:
        csrf_token: { type: string }
        user:
          type: object
          required: [id, is_new]
          properties:
            id: { type: string, format: uuid }
            is_new: { type: boolean }
    Profile:
      type: object
      required:
        [id, sex, birth_year, height_cm, weight_kg, body_fat_pct, goal, experience_level, plan_tier]
      properties:
        id: { type: string }
        sex: { type: string, enum: [male, female, other] }
        birth_year: { type: integer }
        height_cm: { type: number }
        weight_kg: { type: number }
        body_fat_pct: { type: number, nullable: true }
        goal: { type: string, enum: [diet, hypertrophy, strength] }
        experience_level: { type: string, enum: [beginner, intermediate, advanced] }
        days_per_week: { type: integer }
        minutes_per_day: { type: integer }
        plan_tier: { type: string, enum: [free, pro] }
    ProfileUpdate:
      type: object
      properties:
        weight_kg: { type: number }
        body_fat_pct: { type: number }
        goal: { type: string, enum: [diet, hypertrophy, strength] }
        days_per_week: { type: integer }
        minutes_per_day: { type: integer }
        experience_level: { type: string, enum: [beginner, intermediate, advanced] }
    Consent:
      type: object
      required: [type, version, granted]
      properties:
        type: { type: string, example: privacy_policy }
        version: { type: string, example: '2026-07-16' }
        granted: { type: boolean }
    DataExport:
      type: object
      required: [exported_at, profile, consents, data]
      properties:
        exported_at: { type: string, format: date-time }
        profile: { $ref: '#/components/schemas/Profile' }
        consents:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Consent'
              - type: object
                required: [granted_at]
                properties:
                  granted_at: { type: string, format: date-time }
        data:
          type: object
          required: [programs, sessions, estimated_1rm, muscle_weekly_load, calibration_sets, access_audits]
          properties:
            programs: { type: array, items: { type: object, additionalProperties: true } }
            sessions: { type: array, items: { type: object, additionalProperties: true } }
            estimated_1rm: { type: array, items: { type: object, additionalProperties: true } }
            muscle_weekly_load: { type: array, items: { type: object, additionalProperties: true } }
            rir_calibration: { type: object, nullable: true, additionalProperties: true }
            calibration_sets: { type: array, items: { type: object, additionalProperties: true } }
            access_audits: { type: array, items: { type: object, additionalProperties: true } }
...
    Estimated1RM:
      type: object
      required: [exercise_id, e1rm, method, computed_at]
      properties:
        exercise_id: { type: string }
        e1rm: { type: number }
        method: { type: string, example: epley }
        computed_at: { type: string, format: date-time }
...